clean code workinf lfuda

+++ src/rgw/rgw_directory.h
@@ -83,7 +83,7 @@ public:
 	int updateAccessCount(string key, int incr);
 	int setTTL(cache_block *ptr, int seconds);
 	int getHosts(string key, string hosts);
-	int setAvgCacheWeight(size_t avg_weight);
+	int setAvgCacheWeight(int64_t avg_weight);
 	int getAvgCacheWeight(string endpoint);
 private:
 	string buildIndex(cache_block *ptr);
